Move setup command handling to its own msetup.py file
